+++ src/piholecombinelist/__init__.py
@@ -1,7 +1,7 @@
 """Pi-hole Combined Blocklist Generator."""
-# v1.8.0
+# v1.8.1
 
-__version__ = "1.8.0"
+__version__ = "1.8.1"
 
 from .combiner import ListCombiner
 from .database import Database